refactor(openmetrics): split open-metric-server child into server/ + extract pure assembleMetrics
