+++ packages/xo-server-openmetrics/src/open-metric-server.mts
@@ -12,54 +12,22 @@
  * 1. Closing the HTTP server (stops accepting new connections)
  * 2. Disconnecting from the parent IPC channel
  * 3. Letting the event loop drain naturally
+ *
+ * The heavy lifting lives in the `./server/*` modules: request/response IPC
+ * correlation in `request-client.mjs`, RRD fetching in `rrd-fetcher.mjs`, and
+ * metrics gathering/assembly in `collect-metrics.mjs`. This file is the slim
+ * entry: process lifecycle, IPC routing, and the HTTP `/metrics` endpoint.
  */
 
 import { createServer, type Server } from 'node:http'
-import { Agent, request as undiciRequest } from 'undici'
-import { asyncEach } from '@vates/async-each'
 import { createLogger } from '@xen-orchestra/log'
 
-import { parseRrdResponse, type ParsedRrdData } from './rrd-parser.mjs'
-import {
-  formatAllPoolsToOpenMetrics,
-  formatHostStatusMetrics,
-  formatHostUptimeMetrics,
-  formatSrMetrics,
-  formatToOpenMetrics,
-  formatVdiMetrics,
-  formatVmStatusMetrics,
-  formatVmUptimeMetrics,
-  formatXoMetrics,
-  formatXostorAlarmsMetrics,
-  formatXostorClusterMetrics,
-  formatXostorSmartMetrics,
-  formatXostorUpdatesMetrics,
-  type XoMetricsData,
-  type XostorAlarmsPayload,
-  type XostorPayload,
-  type XostorSmartPayload,
-  type XostorUpdatesPayload,
-} from './openmetric-formatter.mjs'
-import type { HostCredentials, IpcMessage, ServerConfiguration } from './types/ipc.mjs'
-import type {
-  HostStatusPayload,
-  SrDataPayload,
-  VdiDataPayload,
-  VmStatusPayload,
-  XapiCredentialsPayload,
-} from './types/domain.mjs'
+import type { IpcMessage, ServerConfiguration } from './types/ipc.mjs'
+import { collectMetrics } from './server/collect-metrics.mjs'
+import { httpsAgent } from './server/rrd-fetcher.mjs'
+import { handleParentResponse, rejectAllPending, sendToParent } from './server/request-client.mjs'
 
 const logger = createLogger('xo:xo-server-openmetrics:child')
-
-// ============================================================================
-// Types
-// ============================================================================
-
-interface PendingRequest<T> {
-  resolve: (value: T) => void
-  reject: (error: Error) => void
-  timer: ReturnType<typeof setTimeout>
-}
 
 // ============================================================================
 // State
@@ -67,34 +35,11 @@ interface PendingRequest<T> {
 
 let server: Server | undefined
 let configuration: ServerConfiguration | undefined
-const pendingRequests = new Map<string, PendingRequest<unknown>>()
-let requestIdCounter = 0
 let isShuttingDown = false
-
-// ============================================================================
-// Constants
-// ============================================================================
-
-const IPC_REQUEST_TIMEOUT_MS = 30_000
-const RRD_FETCH_TIMEOUT_MS = 10_000
-const RRD_FETCH_CONCURRENCY = 5 // Limit parallel RRD requests to avoid overwhelming XAPI
-
-// HTTPS agent that accepts self-signed certificates (common in XenServer)
-const httpsAgent = new Agent({
-  connect: {
-    rejectUnauthorized: false,
-  },
-})
 
 // ============================================================================
 // IPC Communication
 // ============================================================================
-
-function sendToParent(message: IpcMessage): void {
-  if (process.send !== undefined && process.connected) {
-    process.send(message)
-  }
-}
 
 function handleParentMessage(rawMessage: unknown): void {
   const message = rawMessage as IpcMessage
@@ -118,7 +63,7 @@ function handleParentMessage(rawMessage: unknown): void {
     case 'XOSTOR_ALARMS':
     case 'XOSTOR_SMART':
     case 'XOSTOR_UPDATES':
-      resolvePendingRequest(message)
+      handleParentResponse(message)
       break
 
     default:
@@ -151,27 +96,9 @@ async function handleShutdown(): Promise<void> {
   await cleanup()
 }
 
-function resolvePendingRequest(message: IpcMessage): void {
-  const requestId = message.requestId
-  if (requestId === undefined) {
-    return
-  }
-
-  const pending = pendingRequests.get(requestId)
-  if (pending !== undefined) {
-    clearTimeout(pending.timer)
-    pendingRequests.delete(requestId)
-    pending.resolve(message.payload)
-  }
-}
-
 async function cleanup(): Promise<void> {
   // Cancel all pending requests
-  for (const pending of pendingRequests.values()) {
-    clearTimeout(pending.timer)
-    pending.reject(new Error('Server shutting down'))
-  }
-  pendingRequests.clear()
+  rejectAllPending()
 
   // Close the HTTP server
   if (server !== undefined) {
@@ -186,485 +113,6 @@ async function cleanup(): Promise<void> {
   if (process.connected) {
     process.disconnect()
   }
-}
-
-// ============================================================================
-// XAPI Credentials Request
-// ============================================================================
-
-async function requestXapiCredentials(): Promise<XapiCredentialsPayload> {
-  const requestId = `creds-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for XAPI credentials from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as XapiCredentialsPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_XAPI_CREDENTIALS', requestId })
-  })
-}
-
-async function requestSrData(): Promise<SrDataPayload> {
-  const requestId = `sr-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for SR data from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as SrDataPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_SR_DATA', requestId })
-  })
-}
-
-async function requestVdiData(): Promise<VdiDataPayload> {
-  const requestId = `vdi-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for VDI data from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as VdiDataPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_VDI_DATA', requestId })
-  })
-}
-
-async function requestHostStatusData(): Promise<HostStatusPayload> {
-  const requestId = `host-status-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for host status data from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as HostStatusPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_HOST_STATUS', requestId })
-  })
-}
-
-async function requestVmStatusData(): Promise<VmStatusPayload> {
-  const requestId = `vm-status-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for VM status data from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as VmStatusPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_VM_STATUS', requestId })
-  })
-}
-
-async function requestXoMetrics(): Promise<XoMetricsData> {
-  const requestId = `xo-metrics-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for XO metrics from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as XoMetricsData),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_XO_METRICS', requestId })
-  })
-}
-
-async function requestXostorData(): Promise<XostorPayload> {
-  const requestId = `xostor-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for XOSTOR data from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as XostorPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_XOSTOR_DATA', requestId })
-  })
-}
-
-async function requestXostorAlarms(): Promise<XostorAlarmsPayload> {
-  const requestId = `xostor-alarms-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for XOSTOR alarms from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as XostorAlarmsPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_XOSTOR_ALARMS', requestId })
-  })
-}
-
-async function requestXostorSmart(): Promise<XostorSmartPayload> {
-  const requestId = `xostor-smart-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for XOSTOR SMART data from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as XostorSmartPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_XOSTOR_SMART', requestId })
-  })
-}
-
-async function requestXostorUpdates(): Promise<XostorUpdatesPayload> {
-  const requestId = `xostor-updates-${++requestIdCounter}`
-
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => {
-      pendingRequests.delete(requestId)
-      reject(new Error('Timeout waiting for XOSTOR updates from parent'))
-    }, IPC_REQUEST_TIMEOUT_MS)
-
-    pendingRequests.set(requestId, {
-      resolve: value => resolve(value as XostorUpdatesPayload),
-      reject,
-      timer,
-    })
-
-    sendToParent({ type: 'GET_XOSTOR_UPDATES', requestId })
-  })
-}
-
-// ============================================================================
-// RRD Data Fetching
-// ============================================================================
-
-/**
- * Fetch and parse RRD data from a single host.
- *
- * Each host in a XCP-ng/XenServer pool serves RRD data for:
- * - The host itself
- * - All VMs running on that host
- *
- * @param host - Host credentials including address, sessionId, poolId, labels
- * @returns ParsedRrdData or null on error
- */
-async function fetchRrdFromHost(host: HostCredentials): Promise<ParsedRrdData | null> {
-  const { hostAddress, hostLabel, sessionId, poolId, poolLabel, protocol } = host
-
-  // Calculate start time: current time minus 2 intervals (to ensure we get recent data)
-  const now = Math.floor(Date.now() / 1000)
-  const interval = 60
-  const start = now - 2 * interval
-
-  // Build RRD URL with query parameters
-  // cf=AVERAGE is required (cf=LAST is not supported by XenServer rrd_updates endpoint)
-  // Using start parameter to get recent data points
-  const baseUrl = `${protocol}//${hostAddress}`
-  const url = new URL('/rrd_updates', baseUrl)
-  url.searchParams.set('session_id', sessionId)
-  url.searchParams.set('cf', 'AVERAGE')
-  url.searchParams.set('interval', String(interval))
-  url.searchParams.set('start', String(start))
-  url.searchParams.set('host', 'true')
-  url.searchParams.set('json', 'true')
-
-  try {
-    const response = await undiciRequest(url, {
-      method: 'GET',
-      dispatcher: httpsAgent,
-      headersTimeout: RRD_FETCH_TIMEOUT_MS,
-      bodyTimeout: RRD_FETCH_TIMEOUT_MS,
-    })
-
-    if (response.statusCode !== 200) {
-      // Drain the body to release the connection
-      await response.body.text()
-      logger.warn('RRD fetch failed', { hostLabel, poolLabel, statusCode: response.statusCode })
-      return null
-    }
-
-    const text = await response.body.text()
-
-    try {
-      // Parse RRD response using the dedicated parser (handles JSON5 fallback)
-      return parseRrdResponse(text, poolId)
-    } catch (parseError) {
-      logger.warn('RRD parse failed', { hostLabel, poolLabel, error: parseError })
-      return null
-    }
-  } catch (error) {
-    logger.warn('RRD fetch error', { hostLabel, poolLabel, error, url })
-    return null
-  }
-}
-
-// ============================================================================
-// Metrics Collection
-// ============================================================================
-
-/**
- * Collect metrics from all hosts in all connected pools.
- *
- * Fetches RRD data from each host with bounded concurrency,
- * parses and transforms to OpenMetrics format.
- * Also collects SR capacity metrics from XO objects.
- *
- * @returns OpenMetrics-formatted string
- */
-/**
- * Wrap an XOSTOR IPC request so any failure logs a warning and yields the
- * supplied fallback. Keeps `/metrics` responding even when one collector
- * misbehaves.
- */
-function safeXostorRequest<T>(promise: Promise<T>, label: string, empty: T): Promise<T> {
-  return promise.catch((err: unknown) => {
-    logger.warn(`XOSTOR ${label} request failed; emitting empty payload`, { error: err })
-    return empty
-  })
-}
-
-async function collectMetrics(): Promise<string> {
-  const [
-    credentials,
-    srData,
-    vdiData,
-    hostStatusData,
-    vmStatusData,
-    xoMetricsData,
-    xostorData,
-    xostorAlarms,
-    xostorSmart,
-    xostorUpdates,
-  ] = await Promise.all([
-    requestXapiCredentials(),
-    requestSrData(),
-    requestVdiData(),
-    requestHostStatusData(),
-    requestVmStatusData(),
-    requestXoMetrics(),
-    safeXostorRequest(requestXostorData(), 'data', { clusters: [] } as XostorPayload),
-    safeXostorRequest(requestXostorAlarms(), 'alarms', { clusters: [] } as XostorAlarmsPayload),
-    safeXostorRequest(requestXostorSmart(), 'SMART', { hosts: [] } as XostorSmartPayload),
-    safeXostorRequest(requestXostorUpdates(), 'updates', { hosts: [] } as XostorUpdatesPayload),
-  ])
-
-  logger.debug('Collecting metrics', {
-    hostCount: credentials.hosts.length,
-    srCount: srData.srs.length,
-    vdiCount: vdiData.vdis.length,
-    hostStatusCount: hostStatusData.hosts.length,
-    vmStatusCount: vmStatusData.vms.length,
-    xostorClusterCount: xostorData.clusters.length,
-    xostorAlarmClusterCount: xostorAlarms.clusters.length,
-    xostorSmartHostCount: xostorSmart.hosts.length,
-    xostorUpdatesHostCount: xostorUpdates.hosts.length,
-  })
-
-  if (credentials.hosts.length === 0) {
-    return '# No connected hosts\n# EOF'
-  }
-
-  // Collect parsed RRD data from all hosts
-  const rrdDataList: ParsedRrdData[] = []
-  try {
-    await asyncEach(
-      credentials.hosts,
-      async host => {
-        const rrdData = await fetchRrdFromHost(host)
-        if (rrdData !== null) {
-          logger.debug('RRD data fetched', {
-            hostLabel: host.hostLabel,
-            poolLabel: host.poolLabel,
-            metricCount: rrdData.metrics.length,
-          })
-          rrdDataList.push(rrdData)
-        }
-      },
-      { concurrency: RRD_FETCH_CONCURRENCY, stopOnError: false }
-    )
-  } catch (error) {
-    logger.warn('Error collecting RRD metrics', { error })
-  }
-
-  // Build pool connection metrics (deduplicate pools from hosts)
-  const poolMetrics: string[] = []
-  poolMetrics.push('# HELP xcp_pool_connected Indicates if a pool is connected (1) or not (0)')
-  poolMetrics.push('# TYPE xcp_pool_connected gauge')
-
-  const seenPools = new Set<string>()
-  for (const host of credentials.hosts) {
-    if (!seenPools.has(host.poolId)) {
-      seenPools.add(host.poolId)
-      poolMetrics.push(`xcp_pool_connected{pool_id="${host.poolId}",pool_name="${host.poolLabel}"} 1`)
-    }
-  }
-
-  // Format all RRD data to OpenMetrics
-  logger.debug('Formatting RRD data', {
-    hostCount: rrdDataList.length,
-    totalMetrics: rrdDataList.reduce((sum, d) => sum + d.metrics.length, 0),
-  })
-  const rrdMetrics = formatAllPoolsToOpenMetrics(rrdDataList, credentials)
-  logger.debug('Formatted metrics', { outputLength: rrdMetrics.length })
-
-  // Format SR capacity metrics
-  const srMetrics = formatSrMetrics(srData.srs)
-  const srMetricsOutput = srMetrics.length > 0 ? formatToOpenMetrics(srMetrics) : ''
-  logger.debug('Formatted SR metrics', { srCount: srMetrics.length })
-
-  // Format VDI disk size metrics
-  const vdiMetrics = formatVdiMetrics(vdiData.vdis)
-  const vdiMetricsOutput = vdiMetrics.length > 0 ? formatToOpenMetrics(vdiMetrics) : ''
-  logger.debug('Formatted VDI metrics', { vdiCount: vdiMetrics.length })
-
-  // Format host status metrics
-  const hostStatusMetrics = formatHostStatusMetrics(hostStatusData.hosts)
-  const hostStatusOutput = hostStatusMetrics.length > 0 ? formatToOpenMetrics(hostStatusMetrics) : ''
-  logger.debug('Formatted host status metrics', { hostCount: hostStatusMetrics.length })
-
-  // Format host uptime metrics
-  const uptimeMetrics = formatHostUptimeMetrics(credentials)
-  const uptimeMetricsOutput = uptimeMetrics.length > 0 ? formatToOpenMetrics(uptimeMetrics) : ''
-  logger.debug('Formatted host uptime metrics', { hostCount: uptimeMetrics.length })
-
-  // Format VM status metrics
-  const vmStatusMetrics = formatVmStatusMetrics(vmStatusData.vms)
-  const vmStatusOutput = vmStatusMetrics.length > 0 ? formatToOpenMetrics(vmStatusMetrics) : ''
-  logger.debug('Formatted VM status metrics', { vmCount: vmStatusMetrics.length })
-
-  // Format VM uptime metrics
-  const vmUptimeMetrics = formatVmUptimeMetrics(credentials)
-  const vmUptimeOutput = vmUptimeMetrics.length > 0 ? formatToOpenMetrics(vmUptimeMetrics) : ''
-  logger.debug('Formatted VM uptime metrics', { vmCount: vmUptimeMetrics.length })
-
-  // Format XO management plane metrics
-  const xoMetrics = formatXoMetrics(xoMetricsData)
-  const xoMetricsOutput = xoMetrics.length > 0 ? formatToOpenMetrics(xoMetrics) : ''
-  logger.debug('Formatted XO metrics', { count: xoMetrics.length })
-
-  // Format XOSTOR cluster metrics
-  const xostorMetrics = formatXostorClusterMetrics(xostorData)
-  const xostorMetricsOutput = xostorMetrics.length > 0 ? formatToOpenMetrics(xostorMetrics) : ''
-  logger.debug('Formatted XOSTOR metrics', { count: xostorMetrics.length })
-
-  // Format XOSTOR alarm metrics
-  const xostorAlarmsMetrics = formatXostorAlarmsMetrics(xostorAlarms)
-  const xostorAlarmsOutput = xostorAlarmsMetrics.length > 0 ? formatToOpenMetrics(xostorAlarmsMetrics) : ''
-  logger.debug('Formatted XOSTOR alarm metrics', { count: xostorAlarmsMetrics.length })
-
-  // Format XOSTOR SMART metrics
-  const xostorSmartMetrics = formatXostorSmartMetrics(xostorSmart)
-  const xostorSmartOutput = xostorSmartMetrics.length > 0 ? formatToOpenMetrics(xostorSmartMetrics) : ''
-  logger.debug('Formatted XOSTOR SMART metrics', { count: xostorSmartMetrics.length })
-
-  // Format XOSTOR pending-updates metrics
-  const xostorUpdatesMetrics = formatXostorUpdatesMetrics(xostorUpdates)
-  const xostorUpdatesOutput = xostorUpdatesMetrics.length > 0 ? formatToOpenMetrics(xostorUpdatesMetrics) : ''
-  logger.debug('Formatted XOSTOR update metrics', { count: xostorUpdatesMetrics.length })
-
-  // Combine pool metrics with RRD metrics, SR metrics, host status metrics, uptime metrics, and XO metrics
-  // Remove the # EOF from rrdMetrics if present (we'll add our own)
-  const rrdMetricsWithoutEof = rrdMetrics.replace(/\n# EOF$/, '')
-
-  const allMetricsSections = [poolMetrics.join('\n')]
-
-  if (rrdMetricsWithoutEof !== '') {
-    allMetricsSections.push(rrdMetricsWithoutEof)
-  }
-
-  if (srMetricsOutput !== '') {
-    allMetricsSections.push(srMetricsOutput)
-  }
-
-  if (vdiMetricsOutput !== '') {
-    allMetricsSections.push(vdiMetricsOutput)
-  }
-
-  if (hostStatusOutput !== '') {
-    allMetricsSections.push(hostStatusOutput)
-  }
-
-  if (uptimeMetricsOutput !== '') {
-    allMetricsSections.push(uptimeMetricsOutput)
-  }
-
-  if (vmStatusOutput !== '') {
-    allMetricsSections.push(vmStatusOutput)
-  }
-
-  if (vmUptimeOutput !== '') {
-    allMetricsSections.push(vmUptimeOutput)
-  }
-
-  if (xoMetricsOutput !== '') {
-    allMetricsSections.push(xoMetricsOutput)
-  }
-
-  if (xostorMetricsOutput !== '') {
-    allMetricsSections.push(xostorMetricsOutput)
-  }
-
-  if (xostorAlarmsOutput !== '') {
-    allMetricsSections.push(xostorAlarmsOutput)
-  }
-
-  if (xostorSmartOutput !== '') {
-    allMetricsSections.push(xostorSmartOutput)
-  }
-
-  if (xostorUpdatesOutput !== '') {
-    allMetricsSections.push(xostorUpdatesOutput)
-  }
-
-  return allMetricsSections.join('\n') + '\n# EOF'
 }
 
 // ============================================================================
